gotta update LDLib and dump or update registrate

+++ src/main/java/com/gregtechceu/gtceu/api/data/chemical/material/properties/FluidPipeProperties.java
@@ -24,11 +24,9 @@ public class FluidPipeProperties implements IMaterialProperty<FluidPipePropertie
      */
     public static final int MAX_PIPE_CHANNELS = 9;
 
-    @Getter
-    @Setter
-    private long throughput;
-    @Getter
-    @Setter
+    @Getter @Setter
+    private int throughput;
+    @Getter @Setter
     private int channels;
     @Getter
     @Setter
@@ -45,8 +43,8 @@ public class FluidPipeProperties implements IMaterialProperty<FluidPipePropertie
 
     private final Object2BooleanMap<FluidAttribute> containmentPredicate = new Object2BooleanOpenHashMap<>();
 
-    public FluidPipeProperties(int maxFluidTemperature, long throughput, boolean gasProof, boolean acidProof,
-                               boolean cryoProof, boolean plasmaProof, int channels) {
+
+    public FluidPipeProperties(int maxFluidTemperature, int throughput, boolean gasProof, boolean acidProof, boolean cryoProof, boolean plasmaProof, int channels) {
         this.maxFluidTemperature = maxFluidTemperature;
         this.throughput = throughput;
         this.gasProof = gasProof;
@@ -59,8 +57,7 @@ public class FluidPipeProperties implements IMaterialProperty<FluidPipePropertie
     /**
      * Default property constructor.
      */
-    public FluidPipeProperties(int maxFluidTemperature, long throughput, boolean gasProof, boolean acidProof,
-                               boolean cryoProof, boolean plasmaProof) {
+    public FluidPipeProperties(int maxFluidTemperature, int throughput, boolean gasProof, boolean acidProof, boolean cryoProof, boolean plasmaProof) {
         this(maxFluidTemperature, throughput, gasProof, acidProof, cryoProof, plasmaProof, 1);
     }
